+++ tests/testthat/test-updateGeneralSettings.R
@@ -13,14 +13,35 @@ sapply(studies, function(study) {
   test_that("Update a general parameter", {
     
     # year-by-year
-    expect_true(getOption("antares")$parameters$general$`year-by-year`)
-    updateGeneralSettings(year.by.year = FALSE)
-    expect_false(getOption("antares")$parameters$general$`year-by-year`)
+    current_value <- getOption("antares")$parameters$general$`year-by-year`
+    updateGeneralSettings(year.by.year = !current_value, opts = opts)
+    new_value <- getOption("antares")$parameters$general$`year-by-year`
+    if (current_value) {
+      expect_false(new_value)
+    } else {
+      expect_true(new_value)
+    }
     
     # geographic-trimming
-    expect_true(getOption("antares")$parameters$general$`geographic-trimming`)
-    updateGeneralSettings(geographic.trimming = FALSE)
-    expect_false(getOption("antares")$parameters$general$`geographic-trimming`)
+    current_value <- getOption("antares")$parameters$general$`geographic-trimming`
+    updateGeneralSettings(geographic.trimming = !current_value, opts = opts)
+    new_value <- getOption("antares")$parameters$general$`geographic-trimming`
+    if (current_value) {
+      expect_false(new_value)
+    } else {
+      expect_true(new_value)
+    }
+    
+    # thematic-trimming
+    current_value <- getOption("antares")$parameters$general$`thematic-trimming`
+    updateGeneralSettings(thematic.trimming = !current_value, opts = opts)
+    new_value <- getOption("antares")$parameters$general$`thematic-trimming`
+    if (current_value) {
+      expect_false(new_value)
+    } else {
+      expect_true(new_value)
+    }
+    
   })
   
   # remove temporary study
@@ -38,7 +59,7 @@ test_that("updateGeneralSettings() : check appearance of property custom-scenari
   
   # custom-scenario (logical)
   expect_false(getOption("antares")$parameters$general$`custom-scenario`)
-  updateGeneralSettings(custom.scenario = TRUE)
+  updateGeneralSettings(custom.scenario = TRUE, opts = opts)
   expect_true(getOption("antares")$parameters$general$`custom-scenario`)
   # check lower case for a logical value
   lines_generaldata <- readLines(file.path(opts$studyPath, "settings", "generaldata.ini"))